Remove a test case.
When dependent qualified types are supported, this runs into
a different bug, #1913, where we generate an infinitely deep
structure. The reduced template here doesn't really make any
sense and wouldn't be encountered in the wild quite like this.

Previously we generated opaque types for this case, so we didn't
run into that bug. We now generate higher-fidelity types and therefore
hit the problem.

I can't see any way to salvage the fundamentals of this test case
without the recursive problem, so I propose to remove it.